move residual block to nn module
